app: restructure ui navigation
