+++ app/app/src/main/java/de/stustanet/stustapay/ui/nav/NavDest.kt
@@ -1,0 +1,7 @@
+package de.stustanet.stustapay.ui.nav
+
+
+data class NavDest(
+    val route: String,
+    val showNavbar: Boolean = true
+)
