+++ app/app/src/main/java/de/stustanet/stustapay/ui/nav/NavPanel.kt
@@ -1,4 +1,4 @@
-package de.stustanet.stustapay.ui
+package de.stustanet.stustapay.ui.nav
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -20,8 +20,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
-import de.stustanet.stustapay.ui.chipstatus.ChipStatusViewModel
+import de.stustanet.stustapay.ui.root.RootNavDests
 
 data class NavMenuItem(
     val icon: ImageVector,
