+++ app/app/src/main/java/de/stustanet/stustapay/util/SysUiController.kt
@@ -1,0 +1,12 @@
+package de.stustanet.stustapay.util
+
+
+/**
+ * Something that can hide and show menu and title bars.
+ *
+ * Used to call ui hiding/showing on the main activity.
+ */
+interface SysUiController {
+    fun hideSystemUI()
+    fun showSystemUI()
+}
